ref: merge util and utils
